requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
flask_cors